typeにおけるUIの切り替え演出の追加と連続切り替え対策した

+++ Object/Bullet/PlayerBullet/PlayerDefenceBullet.cpp
@@ -37,7 +37,7 @@ void PlayerDefenceBullet::SetDirection(float degree)
 {
 	angle_deg = degree;
 	const float speed = 1500.0f;
-	float rad = degree * DX_PI / 180.0f;
+	float rad = static_cast<float>(degree * DX_PI / 180.0f);
 	velocity.x = speed * std::sin(rad);
 	velocity.y = -speed * std::cos(rad);
 }
